requirements.txt
# ядро приложения
flask==3.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1

# для сессий и кэша
flask-caching==2.1.0

# работа с временем
python-dateutil==2.8.2

# запросы к API
requests==2.31.0
httpx==0.25.2

# конфиг
python-dotenv==1.0.0

# база данных (sqlite уже встроена, но для миграций)
alembic==1.13.1

# утилиты
uuid7==0.1.0

# для десктопного приложения
pywebview==4.4.1
pyinstaller==6.3.0
